Feat: 검색 ui 와 검색어 추천 알고리즘 연결
